+++ R/survival_coxph_pairwise.R
@@ -120,7 +120,7 @@ s_coxph_pairwise <- function(df,
 a_coxph_pairwise <- make_afun(
   s_coxph_pairwise,
   .indent_mods = c(pvalue = 0L, hr = 0L, hr_ci = 1L),
-  .formats = c(pvalue = "x.xxxx | (<0.0001)", hr = "xx.xxxx", hr_ci = "(xx.xxxx, xx.xxxx)")
+  .formats = c(pvalue = "x.xxxx | (<0.0001)", hr = "xx.xx", hr_ci = "(xx.xx, xx.xx)")
 )