+++ NEWS.md
@@ -1,4 +1,5 @@
 # tern 0.7.3
+* Updated default formats of hr and hr_ci in `a_coxph_pairwise` and median in `s_surv_time` to align with STREAM. 
 * Updated `g_km` to now display hazard ratio and its confidence interval to two decimal places as done in STREAM.
 * Updated default position of hazard ratio table in `g_km` to stay
 on the left bottom corner but above x-axis for greater legibility.